libexpr: Lessen const bound on Value::listElems() const
I do not like this being a requirement, but it is necessary to make the
C api work with the least amount of ABI changes.

listElems() const returns a constant pointer to N constant pointers of
Value. This bound is lessened (within defined behavior) to a constant
pointer to N *mutable* pointers of Value. The upstream reasoning behind
this is reasonable enough, though I strongly distrust it.

It might be better to drop the `const` off of the C ABI instead.

Change-Id: I8e750b086c1b8d13972aa748ce8f6139f6bcbadb